requirements.txt
streamlit==1.32.0
langchain==0.1.14
langchain-community==0.0.32
langchain-text-splitters==0.0.1
faiss-cpu==1.7.4
pypdf==4.2.0
sentence-transformers==2.6.1
google-generativeai==0.8.3